+++ src/biz/bokhorst/xprivacy/PrivacyManager.java
@@ -1070,6 +1070,8 @@ public class PrivacyManager {
 
 		// 1 degree ~ 111111 m
 		// 1 m ~ 0,000009 degrees
+		if (location == null)
+			location = new Location(cDeface);
 		location.setLatitude(Float.parseFloat(sLat) + (Math.random() * 2.0 - 1.0) * location.getAccuracy() * 9e-6);
 		location.setLongitude(Float.parseFloat(sLon) + (Math.random() * 2.0 - 1.0) * location.getAccuracy() * 9e-6);
 		location.setAltitude(Float.parseFloat(sAlt) + (Math.random() * 2.0 - 1.0) * location.getAccuracy());