+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Changelog
 **Next release**
 
 * Updated some texts
+* Added restrictions for [Google Maps API v2](http://developer.android.com/reference/com/google/android/gms/maps/GoogleMap.html) ([issue](/../../issues/1807))
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)